requirements.txt
# IGotYou Python Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
fastapi
uvicorn[standard]
pydantic

# AI & Agent
google-adk
google-genai

# Google Services
googlemaps

# Utilities
python-dotenv
